+++ config/styleguide.config.js
@@ -117,11 +117,12 @@ module.exports = {
   template: styleguidistEnv === 'production' ? productionTemplate : devTemplate,
   assetsDir: path.resolve('docs/assets/'),
   styleguideDir: path.resolve(`styleguide/${styleguidistEnv}`),
-  require: [
-    '@tds/core-css-reset/dist/index.css',
-    path.resolve('docs/scss/styleguide.scss'),
-    path.resolve('docs/setup/tds-core-globals.js'),
-  ].concat(styleguidistEnv === 'e2e' ? path.resolve('docs/scss/e2e.css') : []),
+  require: ['babel-polyfill', path.resolve('docs/setup/tds-core-globals.js')].concat(
+    styleguidistEnv === 'e2e' ? path.resolve('docs/scss/e2e.css') : []
+  ),
+  context: {
+    moment: 'moment',
+  },
   styleguideComponents: {
     Editor: path.resolve('docs/components/overrides/Editor/Editor'),
     Logo: path.resolve('docs/components/custom/Logo/Logo'),