+++ packages/DatePicker/rollup.config.js
@@ -1,0 +1,16 @@
+import postcss from 'rollup-plugin-postcss'
+
+import configure from '../../config/rollup.config'
+import { dependencies } from './package.json'
+
+export default configure({
+  input: './DatePicker.jsx',
+  dependencies,
+  css: false,
+  plugins: [
+    postcss({
+      extract: './dist/index.css',
+      sourceMap: false,
+    }),
+  ],
+})