+++ packages/Progress/index.cjs.js
@@ -1,4 +1,3 @@
-require('./dist/index.css')
 const Progress = require('./dist/index.cjs')
 
 module.exports = Progress