+++ packages/ToggleSwitch/__tests__/ToggleSwitch.spec.jsx
@@ -11,11 +11,11 @@ describe('ToggleSwitch', () => {
     name: 'name',
     value: 'value',
     toolTipCopy: 'en',
-    spinnerLabel: 'Request is processing.',
     onClick: mockOnClick,
   }
 
   const doShallow = props => shallow(<ToggleSwitch {...defaultProps} {...props} />)
+  const doMount = props => mount(<ToggleSwitch {...defaultProps} {...props} />)
 
   it('renders', () => {
     const toggleSwitch = mount(<ToggleSwitch {...defaultProps} />)
@@ -86,47 +86,10 @@ describe('ToggleSwitch', () => {
     ).toEqual(true)
   })
 
-  it('internal state should toggle checked when the input is changed by default', () => {
-    const toggleSwitch = doShallow({ checked: true, onClick: mockOnClick })
-
-    const mockEvent = { persist: () => {} }
-
-    toggleSwitch
-      .find(`#${defaultProps.id}`)
-      .at(0)
-      .simulate('click', mockEvent)
-
-    expect(
-      toggleSwitch
-        .find(`#${defaultProps.id}`)
-        .at(0)
-        .prop('aria-checked')
-    ).toEqual(true)
-
-    expect(
-      toggleSwitch
-        .find(`[data-testid="${defaultProps.id}-switch"]`)
-        .at(0)
-        .prop('aria-checked')
-    ).toEqual(true)
-  })
-
-  it('should override internal state change when onClick handler is provided ', () => {
-    const mockEvent = { persist: () => {} }
-    const toggleSwitch = doShallow({ checked: false, onClick: mockOnClick })
-
-    toggleSwitch
-      .find(`#${defaultProps.id}`)
-      .at(0)
-      .simulate('click', mockEvent)
-
-    // Checked state should remain false instead of toggling
-    expect(
-      toggleSwitch
-        .find(`#${defaultProps.id}`)
-        .at(0)
-        .prop('aria-checked')
-    ).toEqual(false)
+  it('aria-checked state should match checked prop', () => {
+    const toggleSwitch = doMount({
+      checked: false,
+    })
 
     expect(
       toggleSwitch
@@ -135,8 +98,13 @@ describe('ToggleSwitch', () => {
         .prop('aria-checked')
     ).toEqual(false)
 
-    // Custom onChange handler should get event delgated to it instead
-    expect(mockOnClick).toHaveBeenCalledWith(mockEvent)
+    toggleSwitch.setProps({ checked: true })
+    expect(
+      toggleSwitch
+        .find(`[data-testid="${defaultProps.id}-switch"]`)
+        .at(0)
+        .prop('aria-checked')
+    ).toEqual(true)
   })
 
   it('should call onFocus handler when provided ', () => {
@@ -172,5 +140,20 @@ describe('ToggleSwitch', () => {
       'data-some-attr',
       'some value'
     )
+  })
+
+  it('should not show spinner when isLoading=false', () => {
+    const toggleSwitch = doShallow({ isLoading: false })
+    const spinner = toggleSwitch.find('Spinner')
+    expect(spinner.prop('spinning')).toEqual(false)
+  })
+
+  it('show spinner when isLoading=true', () => {
+    const toggleSwitch = doShallow({
+      isLoading: true,
+      spinnerLabel: 'Request is processing.',
+    })
+    const spinner = toggleSwitch.find('Spinner')
+    expect(spinner.prop('spinning')).toEqual(true)
   })
 })