+++ packages/Pagination/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.1](https://github.com/telus/tds-community/compare/@tds/community-pagination@2.0.0...@tds/community-pagination@2.0.1) (2019-11-22)
+
+**Note:** Version bump only for package @tds/community-pagination
+
+
+
+
+
 # [2.0.0](https://github.com/telus/tds-community/compare/@tds/community-pagination@1.0.3...@tds/community-pagination@2.0.0) (2019-10-10)